+++ src/components/charts/BarChart.vue
@@ -59,7 +59,14 @@
                 (context.dataset.data[context.dataIndex] ?? 0) !== 0,
               borderRadius: 4,
               padding: 6,
-              formatter: (value) => NumberUtils.formatNumber(value, {})
+              formatter: (value) =>
+                NumberUtils.formatNumber(value, {
+                  decimals: settingStore.numberFormat_decimals.value,
+                  decimalSeparator:
+                    settingStore.numberFormat_decimalSeparator.value,
+                  groupBy: settingStore.numberFormat_groupBy.value,
+                  groupSeparator: settingStore.numberFormat_groupSeparator.value
+                })
             }
           }
         }
@@ -77,6 +84,8 @@
   import { Utils } from '../../utils/utils';
   import { NumberUtils } from '../../utils/numberUtils';
 
+  import { useSettingStore } from '../../stores/settingStore';
+
   export type DataItem = {
     label: string;
     value: number;
@@ -88,6 +97,8 @@
       required: true
     }
   });
+
+  const settingStore = useSettingStore();
 
   const color = Utils.getValueOfCSSVar('--color-primary-element');
   const borderRadius = Utils.getValueOfCSSVar('--border-radius-large').replace(